+++ test/service/test_serialization.py
@@ -1,13 +1,15 @@
 from test.apps.openapi.schema import OpenAPIVersion
+from unittest.mock import ANY
 
 import pytest
 
 import schemathesis
 from schemathesis.runner import from_schema
+from schemathesis.runner.events import InternalError
 from schemathesis.service.serialization import prepare_query, serialize_event, stringify_path_parameters
 
 
-@pytest.mark.operations("multiple_failures")
+@pytest.mark.operations("success", "multiple_failures")
 @pytest.mark.parametrize("openapi_version", (OpenAPIVersion("3.0"),))
 def test_serialize_event(schema_url):
     schema = schemathesis.from_uri(schema_url)
@@ -15,9 +17,33 @@ def test_serialize_event(schema_url):
     next(events)
     next(events)
     event = serialize_event(next(events))
-    assert "interactions" not in event["result"]
-    assert "logs" not in event["result"]
-    assert event["result"]["checks"][0]["example"]["query"] == {"id": ["0"]}
+    assert "interactions" not in event["AfterExecution"]["result"]
+    assert "logs" not in event["AfterExecution"]["result"]
+    assert event["AfterExecution"]["result"]["checks"][0]["example"]["query"] == {"id": ["0"]}
+
+
+@pytest.mark.operations("success")
+@pytest.mark.parametrize("openapi_version", (OpenAPIVersion("3.0"),))
+def test_serialize_interrupted(mocker, schema_url):
+    mocker.patch("schemathesis.runner.impl.solo.SingleThreadRunner._execute_impl", side_effect=KeyboardInterrupt)
+    schema = schemathesis.from_uri(schema_url)
+    events = from_schema(schema).execute()
+    next(events)
+    assert serialize_event(next(events)) == {"Interrupted": None}
+
+
+def test_serialize_internal_error():
+    try:
+        1 / 0
+    except ArithmeticError as exc:
+        event = InternalError.from_exc(exc)
+        assert serialize_event(event) == {
+            "InternalError": {
+                "message": "An internal error happened during a test run",
+                "exception_type": "builtins.ZeroDivisionError",
+                "exception_with_traceback": ANY,
+            }
+        }
 
 
 @pytest.mark.parametrize(