+++ docs/changelog.rst
@@ -4,6 +4,10 @@ Changelog
 `Unreleased`_ - TBD
 -------------------
 
+**Changed**
+
+- Minor changes to the Schemathesis.io integration.
+
 `3.11.3`_ - 2021-12-02
 ----------------------